split out common job code
